Bots/Discord/requirements.txt
paho-mqtt
discord.py